+++ helpers/externals/drive/scopes.ts
@@ -4,7 +4,6 @@ const scopes: TGooGScope[] = [
   'https://www.googleapis.com/auth/drive',
   'https://www.googleapis.com/auth/drive.file',
   'https://www.googleapis.com/auth/drive.metadata',
-  'https://www.googleapis.com/auth/drive.file',
 ];
 
 export default scopes;