+++ helpers/externals/google/api.ts
@@ -1,0 +1,6 @@
+export default {
+  authorize: 'https://accounts.google.com/o/oauth2/v2/auth',
+  generateToken: 'https://oauth2.googleapis.com/token',
+  checkToken: 'https://oauth2.googleapis.com/tokeninfo',
+  revokeToken: 'https://oauth2.googleapis.com/revoke',
+};