+++ routes/first-setup/index.ts
@@ -6,7 +6,6 @@ import policies from './policies';
 import roles from './roles';
 import credentials from './credentials';
 import scopes from './scope';
-import tokens from './token';
 import endpointServer from '@/helpers/express/other-handlers/endpoint-server';
 
 // Router
@@ -17,7 +16,6 @@ router.use('/policies', policies);
 router.use('/roles', roles);
 router.use('/credentials', credentials);
 router.use('/scopes', scopes);
-router.use('/tokens', tokens);
 
 // Respond with all the Endpoints in the Route
 router.get('/endpoints', (req, res) => endpointServer(res, router));