+++ routes/index.ts
@@ -12,6 +12,7 @@ import { okResponse } from '@/helpers/express/response-handlers/2XX-response';
 import firstSetup from './first-setup';
 import policies from './policies';
 import roles from './roles';
+import tokens from './tokens';
 import endpointServer from '@/helpers/express/other-handlers/endpoint-server';
 
 // Router
@@ -21,6 +22,7 @@ const router = express.Router();
 router.use('/setup', [setupCheck, checkSecretPass], firstSetup);
 router.use('/policies', policies);
 router.use('/roles', roles);
+router.use('/tokens', tokens);
 
 // Respond with all the Endpoints in this Route
 router.get('/endpoints', (req, res) => endpointServer(res, router));